requirements.txt
asgiref==3.5.2
boto3==1.24.95
beautifulsoup4==4.9.1
bootstrap4==0.1.0
cryptography==39.0.0
Django==3.2
django-cors-headers==3.11.0
django-environ==0.9.0
django-storages==1.13.1
djangorestframework==3.11.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.17.1
openpyxl==3.1.2
pytz==2020.1
psycopg2-binary==2.9.6
sqlparse==0.4.1
s3transfer==0.6.0
requests==2.26.0
pandas==1.5.2
reportlab==4.0.5
SQLAlchemy==1.3.18
Pillow==9.2.0
